frontend/requirements.txt
streamlit>=1.28.0
pandas>=1.3.0
requests>=2.26.0
python-dotenv>=0.19.0
